♻️  research_forum 1052줄 → 4 모듈 분해 (P0-Q step 2)
변경 이유
- P0-Q 첫 번째 leaf — research_forum.py 는 agents.research.pack 만
  의존하는 가장 안전한 추출 대상. P0-L / P0-P 패턴 재사용.

주요 변경 사항
- ``git mv research_forum.py → research_forum/_legacy.py`` 후
  4 모듈로 단일 commit 분해 (audit doc 의 single-PR 원칙):
  - ``config.py`` (60 줄): ``ResearchForumContext`` + env reader
  - ``prefixes.py`` (33 줄): PREFIX_* 5종 + ``detect_thread_prefix``
  - ``formatters.py`` (~620 줄): pure title/body/comment formatter
    (``derive_research_topic``, ``normalize_thread_title``,
    ``format_research_post_body``, ``format_agent_comment``,
    ``format_thread_markdown_fallback``, ``truncate_for_starter_message``,
    ``split_forum_starter_and_replies`` 및 helper 17 종)
  - ``posting.py`` (~330 줄): Discord async 호출부
    (``ForumPostOutcome`` / ``ForumCommentOutcome`` dataclass,
    ``create_research_post``, ``post_agent_comment``,
    ``chunk_for_discord_message``)
- ``__init__.py`` facade 가 4 모듈에서 공식 API 직접 import.
- ``_legacy.py`` 즉시 제거 (단일 commit 분해).
- 상대 import depth 보정: ``..agents`` → ``...agents``.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- 외부 28+ import 사이트 무회귀: bot.py + tests.
- 다음 step: engineering_team_runtime.py 2153줄.
